+++ CHANGELOG.md
@@ -6,9 +6,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-
 ### Fixed
 - Fix - Empty Message Log issue
+
+### Added
+- Add toggle feature in code if toggle is on then get embedded data and if toggle is off then get response without rag
+- Add prompt for toggle off prompt
+
+### Added
+- Reformat code for adding content type
+
+-Fixed the issue in Metrics & Diagnostics
 
 ### Added
 - Updated tiktoken version due to updated model